Add Dockerfile and .dockerignore files
This commit introduces a Dockerfile which sets up a Rust build environment, compiles the application in release mode and sets up a runtime image. The result is a Docker image that runs our application as the user 'app'. Additionally, a .dockerignore file has been created to exclude the 'target/' directory.